python/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
black==24.8.0
cattrs==23.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.9
distro==1.9.0
dnspython==2.6.1
ell-ai==0.0.13
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
imagekitio==4.1.0
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.1
openai==1.51.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.21
starlette==0.37.2
tokenizers==0.20.0
tomli==2.0.2
tqdm==4.66.5
twilio==9.3.3
typer==0.12.5
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
yarl==1.13.1
